Remove overrides folder
