--- mmv1/overrides/terraform/resource_override.rb
@@ -1,54 +1,0 @@
-# Copyright 2017 Google Inc.
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-require 'overrides/resources'
-require 'provider/terraform/custom_code'
-require 'provider/terraform/docs'
-require 'provider/terraform/examples'
-require 'provider/terraform/virtual_fields'
-
-module Overrides
-  module Terraform
-    # A class to control overridden properties on terraform.yaml in lieu of
-    # values from ResourceName.yaml.
-    class ResourceOverride < Overrides::ResourceOverride
-      def self.attributes
-        []
-      end
-
-      attr_reader(*attributes)
-      attr_reader :description
-
-      def apply(resource)
-        unless description.nil?
-          @description = format_string(:description, @description,
-                                       resource.description)
-        end
-
-        super
-      end
-
-      private
-
-      # Formats the string and potentially uses its old value as part of the new
-      # value. The marker should be in the form `{{name}}` where `name` is the
-      # field being formatted.
-      #
-      # Note: This function only supports the variable with the same name as the
-      # property being updated.
-      def format_string(name, mask, current_value)
-        mask.gsub "{{#{name.id2name}}}", current_value
-      end
-    end
-  end
-end
