+++ mmv1/provider/terraform/validation.rb
@@ -1,4 +1,4 @@
-# Copyright 2017 Google Inc.
+# Copyright 2023 Google Inc.
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -11,13 +11,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
---- !ruby/object:Provider::Terraform::Config
-overrides: !ruby/object:Overrides::ResourceOverrides
-  ResourceWithTerraformOverride: !ruby/object:Overrides::Terraform::ResourceOverride
-    properties:
-      objectTwoFlattened: !ruby/object:Overrides::Terraform::PropertyOverride
-        flatten_object: true
-      objectTwoFlattened.objectTwoNestedObject: !ruby/object:Overrides::Terraform::PropertyOverride
-        update_mask_fields:
-          - 'overrideFoo'
-          - 'nested.overrideBar'
+require 'api/object'
+require 'provider/abstract_core'
+
+module Provider
+  class Terraform < Provider::AbstractCore
+    # Support for schema ValidateFunc functionality.
+    class Validation < Api::Object
+      # Ensures the value matches this regex
+      attr_reader :regex
+      attr_reader :function
+
+      def validate
+        super
+
+        check :regex, type: String
+        check :function, type: String
+      end
+    end
+  end
+end
