+++ mmv1/spec/data/terraform-config.yaml
@@ -12,18 +12,4 @@
 # limitations under the License.
 
 --- !ruby/object:Provider::Terraform::Config
-overrides: !ruby/object:Overrides::ResourceOverrides
-  ResourceWithTerraformOverride: !ruby/object:Overrides::Terraform::ResourceOverride
-    properties:
-      stringRename: !ruby/object:Overrides::Terraform::PropertyOverride
-        name: 'stringRenamed'
-      objectOne.objectOneRename: !ruby/object:Overrides::Terraform::PropertyOverride
-        name: 'objectOneRenamed'
-      objectOne.objectOneFlattenedObject: !ruby/object:Overrides::Terraform::PropertyOverride
-        flatten_object: true
-      objectTwoFlattened: !ruby/object:Overrides::Terraform::PropertyOverride
-        flatten_object: true
-      objectTwoFlattened.objectTwoString: !ruby/object:Overrides::Terraform::PropertyOverride
-        update_mask_fields:
-          - 'overrideFoo'
-          - 'nested.overrideBar'
+files: !ruby/object:Provider::Config::Files
