+++ mmv1/spec/provider_terraform_spec.rb
@@ -25,11 +25,10 @@ describe Provider::Terraform do
     let(:product) { Api::Compiler.new(File.read('spec/data/good-file.yaml')).run }
     let(:parsed) { Provider::Config.parse('spec/data/terraform-config.yaml', product) }
     let(:config) { parsed[1] }
-    let(:override_product) { parsed[0] }
     let(:provider) { Provider::Terraform.new(config, product, 'ga', Time.now) }
     let(:resource) { product.objects[0] }
     let(:override_resource) do
-      override_product.objects.find { |o| o.name == 'ResourceWithTerraformOverride' }
+      product.objects.find { |o| o.name == 'ThirdResource' }
     end
 
     before do
@@ -152,7 +151,7 @@ describe Provider::Terraform do
         is_expected.to eq(
           [
             ['string_one', ['stringOne']],
-            ['string_renamed', ['stringRename']],
+            ['string_renamed', ['stringRenamed']],
             ['object_one', ['objectOne']],
             ['object_two_string', ['overrideFoo', 'nested.overrideBar']],
             [
